refactor(envs): unify futures _get_observation into TorchTradeFuturesLiveEnv
The 4 live futures envs (binance/bitget/bybit/okx) each hand-rolled a byte-identical
_get_observation + _get_portfolio_value (~90-120 lines apiece), so every account_state
fix had to land in 4 copies -- the "fix the instance not the class" trap that made
holding_time (#252) and the liq-distance guard (#253) 4-way sweeps.

Lift one shared copy into a new intermediate base TorchTradeFuturesLiveEnv(TorchTradeLiveEnv);
the 4 futures base envs inherit it and drop their copies. Behaviour-preserving: the shared
body is byte-identical to the old bybit/okx version, and binance/bitget differed only by two
dead locals (cash, entry_price -- assigned, never read) which are removed. Alpaca (spot) is
untouched -- it keeps its own _get_observation (leverage=1, no liquidation, cash-based).

Net -283 LOC. New structural guard test_no_futures_env_reforks_the_shared_observation asserts
all 12 futures classes resolve _get_observation/_get_portfolio_value to the shared base by
identity, so a future re-fork fails loudly instead of silently un-doing the dedup. The
existing AST hold-counter guards already scan the new file (live/** rglob).

Two PRE-EXISTING bugs surfaced during review, logged as separate follow-ups (NOT fixed here --
this PR is behaviour-preserving):
- exposure_pct denominator: Binance's total_wallet_balance excludes unrealized PnL while the
  other 3 map equity (incl uPnL) to the same key -> cross-exchange skew on account_state[0].
- base_features emitted but absent from observation_spec on binance/bitget/bybit (only okx
  declares it) -> check_env_specs fails for those 3.

Reviewed over 3 rounds x 4 agents (test-justification, code-simplifier, pr-test-analyzer,
torchrl-engineer).
